@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 
+import 'core/theming/app_sizer.dart';
+import 'products/presentation/screens/products_screen.dart';
+
 void main() {
   runApp(const MyApp());
 }
@@ -9,13 +12,19 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      home: Container(),
+    return LayoutBuilder(
+      builder: (context, _) {
+        AppSizer.initialize(context);
+        return MaterialApp(
+          debugShowCheckedModeBanner: false,
+          title: 'Flutter Demo',
+          theme: ThemeData(
+            colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+            useMaterial3: true,
+          ),
+          home: const ProductsScreen(),
+        );
+      },
     );
   }
 }
